Extract member creation logic into separate module to re-use it for sharing a WP with a user that does not exist yet
